moderation-service/requirements.txt
flask==3.0.0
flask-limiter==3.5.0
flask-cors==4.0.0
llama-cpp-python==0.2.90
pydantic==2.5.3
python-dotenv==1.0.0
